# internal/server/api/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spivot Server API",
    "version": "0.1.0",
    "description": "Native HTTP API of Spivot Server, the Go backend for the Spivot iOS app, speaking the OpenCaravan protocol.\n\nThis document is generated from the Go source (route table + contract structs) by internal/tools/openapigen; it cannot drift from the implementation. The `x-spivot-auth` extension on each operation records the enforced authentication posture, derived from the same table that wires the middleware."
  },
...
  "paths": {
...
    "/v1/client-apps/enroll": {
      "post": {
        "operationId": "clientAppEnroll",
        "summary": "Enroll a client app with a server-registration invite",
        "tags": [
          "Identity"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientAppEnrollmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientAppEnrollmentResponse"
                }
              }
            }
          }
        },
        "x-spivot-auth": {
          "posture": "public"
        }
      }
    },
    "/v1/sessions": {
      "post": {
        "operationId": "sessionCreate",
        "summary": "Issue a session macaroon",
        "tags": [
          "Identity"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "mutualTLS": []
          }
        ],
        "x-spivot-auth": {
          "posture": "identity"
        }
      }
    },
    "/v1/client-apps/invites": {
      "post": {
        "operationId": "clientAppInviteCreate",
        "summary": "Mint a server-registration invite",
        "tags": [
          "Identity"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientAppInviteCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientAppInviteResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "mutualTLS": []
          }
        ],
        "x-spivot-auth": {
          "posture": "identity"
        }
      },
      "get": {
        "operationId": "clientAppInviteList",
        "summary": "List invites minted by the caller",
        "tags": [
          "Identity"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientAppInviteListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "mutualTLS": []
          }
        ],
        "x-spivot-auth": {
          "posture": "identity"
        }
      }
    },
    "/v1/journeys": {
      "post": {
        "operationId": "journeyCreate",
        "summary": "Create a journey",
        "tags": [
          "Journeys"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "mutualTLS": []
          }
        ],
        "x-spivot-auth": {
          "posture": "identity"
        }
      }
    },
    "/v1/journeys/{id}": {
      "get": {
        "operationId": "journeyGet",
        "summary": "Fetch a journey",
        "tags": [
          "Journeys"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Journey identifier; must match the session macaroon's journey caveat."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "macaroonSession": []
          }
        ],
        "x-spivot-auth": {
          "posture": "session",
          "session_action": "journey.read",
          "journey_path_param": "id"
        }
      }
    },
    "/v1/journeys/{id}/telemetry": {
      "post": {
        "operationId": "telemetryBatchRecord",
        "summary": "Record a telemetry batch",
        "tags": [
          "Journeys"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Journey identifier; must match the session macaroon's journey caveat."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatchRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryBatchResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "macaroonSession": []
          }
        ],
        "x-spivot-auth": {
          "posture": "session",
          "session_action": "telemetry.write",
          "journey_path_param": "id"
        }
      }
    },
    "/v1/journeys/{id}/vehicles": {
      "post": {
        "operationId": "journeyVehicleCreate",
        "summary": "Upload a journey vehicle bundle",
        "tags": [
          "Journey Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Journey identifier; must match the session macaroon's journey caveat."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyVehicleCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyVehicleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "macaroonSession": []
          }
        ],
        "x-spivot-auth": {
          "posture": "session",
          "session_action": "journey.write",
          "journey_path_param": "id"
        }
      },
      "get": {
        "operationId": "journeyVehicleList",
        "summary": "List journey vehicles",
        "tags": [
          "Journey Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Journey identifier; must match the session macaroon's journey caveat."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyVehicleListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "macaroonSession": []
          }
        ],
        "x-spivot-auth": {
          "posture": "session",
          "session_action": "journey.read",
          "journey_path_param": "id"
        }
      }
    },
    "/v1/journeys/{id}/vehicles/{vid}": {
      "get": {
        "operationId": "journeyVehicleGet",
        "summary": "Fetch a journey vehicle",
        "tags": [
          "Journey Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Journey identifier; must match the session macaroon's journey caveat."
          },
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyVehicleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "macaroonSession": []
          }
        ],
        "x-spivot-auth": {
          "posture": "session",
          "session_action": "journey.read",
          "journey_path_param": "id"
        }
      }
    },
    "/v1/journeys/{id}/vehicles/{vid}/acl-revisions": {
      "post": {
        "operationId": "journeyVehicleACLAppend",
        "summary": "Append a signed vehicle ACL revision",
        "tags": [
          "Journey Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Journey identifier; must match the session macaroon's journey caveat."
          },
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleACL"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyVehicleACLRevisionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "macaroonSession": []
          }
        ],
        "x-spivot-auth": {
          "posture": "session",
          "session_action": "journey.write",
          "journey_path_param": "id"
        }
      }
    },
    "/v1/journeys/{id}/vehicles/{vid}/revisions": {
      "post": {
        "operationId": "journeyVehicleRevisionAppend",
        "summary": "Append a signed vehicle metadata revision",
        "tags": [
          "Journey Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Journey identifier; must match the session macaroon's journey caveat."
          },
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vehicle"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyVehicleRevisionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "macaroonSession": []
          }
        ],
        "x-spivot-auth": {
          "posture": "session",
          "session_action": "journey.write",
          "journey_path_param": "id"
        }
      }
    },
...
    "/v1/garages/{id}": {
      "get": {
        "operationId": "garageGet",
        "summary": "Fetch a garage",
        "tags": [
          "Garages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "mutualTLS": []
          }
        ],
        "x-spivot-auth": {
          "posture": "identity"
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "mutualTLS": {
        "type": "mutualTLS",
        "description": "An enrolled client-app certificate, presented directly via mTLS or forwarded by a trusted edge proxy. Obtained from `clientAppEnroll`."
      },
      "macaroonSession": {
        "type": "http",
        "scheme": "macaroon",
        "description": "`Authorization: Macaroon <token>` — a base64url session macaroon minted by `sessionCreate`, carrying user/client-app/action caveats and, for journey-scoped operations, a journey caveat that must match the journey named in the request path."
      }
    },
    "schemas": {
...
      "ClientAppInviteCreateRequest": {
        "description": "ClientAppInviteCreateRequest is the optional POST body. An empty body\n(or no Content-Length) is valid and selects the defaults.",
        "type": "object",
        "properties": {
          "expires_in_seconds": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}